requirements.txt
# AI Trading Research - Dependencies
# Core
requests>=2.31.0
python-dotenv>=1.0.0

# LLM (optional - for report generation)
openai>=1.0.0
google-genai>=1.0.0

# Data & analysis (optional - for advanced technical analysis)
pandas>=2.0.0
numpy>=1.24.0
ta>=0.11.0
